Express router
Configured the express router to handle the 'blogs' routes. First created a routes directory, created blogRoutes file and move the routes related to blogs (starting with '/blogs') there, then imported express and created a router instance. Then imported all the resources required in the request handling methods too. exported the router. Imported the blogRoutes in the app.js then used the 'app.use()' to handle the blog routes and scoped out the blog routes further.
